requirements.txt
python-dotenv==1.0.1
PyPDF2==3.0.1
openai==1.35